pyproject.toml
[tool.poetry]
name = "how-use-airflow"
version = "0.1.0"
description = ""
authors = ["Rodrigo Coelho <rodrigo.coelho@hotmail.com.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
apache-airflow = "^2.7.1"


[tool.poetry.group.dev.dependencies]
loguru = "^0.7.2"
black = "^23.9.1"
isort = "^5.12.0"
autoflake = "^2.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
